create mock request and response. enable actuator
