+++ backend/src/main/java/com/hoppylearn/model/request/DeckRequest.java
@@ -1,0 +1,14 @@
+package com.hoppylearn.model.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class DeckRequest {
+    private String deckName;
+}
