+++ backend/src/main/java/com/hoppylearn/model/response/DeckResponse.java
@@ -1,0 +1,15 @@
+package com.hoppylearn.model.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class DeckResponse {
+    private int id;
+    private String deckName;
+}
